extract clone canvas logic into hook
